Allow uninstalling user shortcuts as administrators (#414)

+++ menuinst/platforms/linux.py
@@ -66,13 +66,17 @@ class LinuxMenu(Menu):
         return (path,)
 
     def remove(self) -> tuple[os.PathLike]:
+        if not Path(self.desktop_entries_location).exists():
+            return tuple()
         for fn in os.listdir(self.desktop_entries_location):
             if fn.startswith(f"{self.render(self.name, slug=True)}_"):
                 # found one shortcut, so don't remove the name from menu
-                return (self.directory_entry_location,)
-        unlink(self.directory_entry_location, missing_ok=True)
+                return tuple()
         self._remove_this_menu()
-        return (self.directory_entry_location,)
+        if self.directory_entry_location.exists():
+            unlink(self.directory_entry_location, missing_ok=True)
+            return (self.directory_entry_location,)
+        return tuple()
 
     @property
     def placeholders(self) -> dict[str, str]:
@@ -112,6 +116,8 @@ class LinuxMenu(Menu):
     #
 
     def _remove_this_menu(self):
+        if not Path(self.menu_config_location).exists():
+            return
         log.debug(
             "Editing %s to remove %s config", self.menu_config_location, self.render(self.name)
         )
@@ -201,12 +207,13 @@ class LinuxMenuItem(MenuItem):
         return self._paths()
 
     def remove(self) -> Iterable[os.PathLike]:
-        paths = self._paths()
+        paths = (path for path in self._paths() if Path(path).is_file())
         self._maybe_register_mime_types(register=False)
-        for path in paths:
-            log.debug("Removing %s", path)
-            unlink(path, missing_ok=True)
-        self._update_desktop_database()
+        if paths:
+            for path in paths:
+                log.debug("Removing %s", path)
+                unlink(path)
+            self._update_desktop_database()
         return paths
 
     def _update_desktop_database(self):
